backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.1.0
authlib==1.3.0
httpx==0.25.2
python-dateutil==2.8.2
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.2.0